Method Each on Query
